requirements.txt
fabric
Django
djangorestframework
